+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * chore: switch to pnpm;
 * chore: replaced husky with dependency-free solution;
 * refactor: abstraction for `forFunctional` to allow use in preact without `preact/compat`;
+* refactor: better types;
 
 ### 1.0.1
 * refactor: getters and mutations simplified;